Updated PHP script in includes/admin/setup.php
Step 4 is now a summary of the data to be saved.
After submitting, the new config is saved.
Then, the page shows the Thanks you section and a link to the home page.

(for the future: a fallback should be added in case of an error in the base url..)